+++ backend/src/main/java/com/HelloRolha/HR/feature/goout/model/dto/GooutTypeCreateReq.java
@@ -1,11 +1,11 @@
-package com.HelloRolha.HR.feature.goout.model.entity;
+package com.HelloRolha.HR.feature.goout.model.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Builder
 @Data
-public class GooutTypeRead {
+public class GooutTypeCreateReq {
     private String name;
     private String detail;
     private Integer maxHoliday;
