+++ backend/src/main/java/com/HelloRolha/HR/feature/goout/model/dto/GooutTypeRead.java
@@ -1,12 +1,11 @@
-package com.HelloRolha.HR.feature.goout.model.entity;
+package com.HelloRolha.HR.feature.goout.model.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Builder
 @Data
-public class GooutTypeUpdateReq {
-    private Integer id;
+public class GooutTypeRead {
     private String name;
     private String detail;
     private Integer maxHoliday;
