+++ backend/src/main/java/com/HelloRolha/HR/feature/goout/model/dto/GooutTypeUpdateReq.java
@@ -1,11 +1,12 @@
-package com.HelloRolha.HR.feature.goout.model.entity;
+package com.HelloRolha.HR.feature.goout.model.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Builder
 @Data
-public class GooutTypeCreateReq {
+public class GooutTypeUpdateReq {
+    private Integer id;
     private String name;
     private String detail;
     private Integer maxHoliday;
